Add hackish support for brightness, contrast and saturation.
Support for hue is there in principle, but would need tweaking,
Due to the very limited range supported by the driver I didn't
manage to do that.
The hardware actually has full support for arbitrary CSC matrices,
but since it is not exposed some hack like this is the best I could
think of.
Patch also fixes vdp_generate_csc_matrix to write the data to
the correct location, previously it would actually write outside
the matrix.

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>