Add XEventLoop to react on window resize, movement,
mapping and unmapping.

It's now possible to move the windwow and resize it,
during replay. It's also possible to unmap and map a window
within a window manager.
The resource intensive XClearWindow was called every frame.
Now it's only called, when some change in dimension and/ or
position was made.
There is still a problem when doing an opaque window move
which has to be fixed. This is using too much resources and
WAF isn't too high when doing that. The only workaround atm is
to disable the opaque move.
Also remember, that video window will always be on top because
of the current sunxi specific implementation.

This commit is based on a fix from vdr-portal.de user johns:
http://www.vdr-portal.de/board18-vdr-hardware/board98-arm-com/p1238936-sunxi-vdpau-wip-ehemals-interlaced-branch/#post1238936

There are still too many missed frames in vdr/softhddevice
when OSD is displayed. This ends up in a stuttering video
in the background, which maybe acceptable atm but not the final goal.