Introduce a thread_control function.
This allows us to control the thread via a task that is pushed into the queue.
A flag is responsible for what happens then.
Reinitiation of the display engine and ending the thread are the two
possibilities at the moment.

Disable video layer when nothing is rendered:
Now we check in the presentation thread, if the video surface changed
and trigger a display init if necessary.