Strip rgba surface handling.
The previous behaviour of the rgba output surface handling was wrong.
We may have a (ring)buffer with output surfaces into which the application
may render. To save recources on sunxi (skipping the G2D blit),
we have to check, if the new input surface (bitmap or output surface)
is already rendered on the destination output surface.
Therefore we add some struct to the rgba surface which archives the
parameters of the last vdp_output_surface_render_output/bitmap_surface
(colors, flags, destination_rect, source_rect) and the unique id of the
rgba surface (which is incremented with every put_bits_native/indexed)
that was rendered into that surface.

Before a new G2D blit is triggered, we can check, if the new rgba is already
rendered in the destination surface, and if so, we simply do nothing, but keeping
the rgba surface dirty.

To keep it more easy than it was before, we display the OSD part of the
output surface (rgba) every time we call do_presentation_queue_display
(each frame), depending if the rgba surface is dirty (RGBA_FLAG_DIRTY).
The cost of resources for this is negligible.

Now it should be possible to do several render_output_surface calls
with differents parameters (e.g. rect size) in sequence without breaking
anything - but blit optimized.